Tweaking completers
