+++ cli/src/main/java/org/arend/frontend/repl/jline/KeywordCompleter.java
@@ -16,7 +16,7 @@ public class KeywordCompleter implements Completer {
       "Prop", "where", "with", "use", "cowith", "elim", "new", "pi", "sigma", "lam", "let", "lets",
       "in", "case", "scase", "data", "class", "record", "module", "meta", "extends",
       "return", "coerce", "instance", "truncated",
-      "lp", "lh", "oo", "suc", "max", "level",
+      "lp", "lh", "oo", "suc", "max", "level", "oo-Type",
       "Set", "Type", "this",
       "eval", "peval"
   );
